Moved navigation from App.js to Navigate.js. Passed userId as context. Fixed recipe id prop.
